build: Enable -Wswitch-enum and -Wswitch-default
This patch makes that work using 1 of 2 options:

1. Add all missing enums to the switch statement
  or
2. Cast the switch argument to a uint to avoid having to do that (mostly
   for GdkEventType).

I even found a bug while doing that: clearing a GtkImage with a surface
did not notify thae surface property.

The reason for enabling this flag even though it is tedious at times is
that it is very useful when adding values to an enum, because it makes
GTK immediately warn about all the switch statements where this enum is
relevant.
And I expect changes to enums to be frequent during the GTK4 development
cycle.

+++ gdk/gdkdevice.c
@@ -1051,7 +1051,7 @@ gdk_device_set_axis_use (GdkDevice   *device,
   info = &g_array_index (device->axes, GdkAxisInfo, index_);
   info->use = use;
 
-  switch (use)
+  switch ((guint) use)
     {
     case GDK_AXIS_X:
     case GDK_AXIS_Y:
@@ -1574,7 +1574,7 @@ _gdk_device_add_axis (GdkDevice   *device,
   axis_info.max_value = max_value;
   axis_info.resolution = resolution;
 
-  switch (use)
+  switch ((guint) use)
     {
     case GDK_AXIS_X:
     case GDK_AXIS_Y:
